requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.7
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
EasyProcess==1.1
entrypoint2==1.1
filelock==3.12.4
fonttools==4.46.0
joblib==1.3.2
kiwisolver==1.4.5
Markdown==3.6
matplotlib==3.8.2
MouseInfo==0.1.3
mss==9.0.1
nltk==3.8.1
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
platformdirs==3.11.0
psycopg2==2.9.9
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3.post1
regex==2023.10.3
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
timedelta==2020.12.3
tqdm==4.66.1
tzdata==2023.3
virtualenv==20.24.6
wordcloud==1.9.3
